[KOGITO-3840] move transportContext to processInstance
