+++ jbpm/process-serialization-protobuf/src/main/resources/org/kie/kogito/serialization/process/protobuf/kogito_types.proto
@@ -28,7 +28,6 @@ message WorkflowContext {
     repeated NodeInstance node_instance = 2;
     repeated NodeInstanceGroup exclusive_group = 3;
     repeated IterationLevel iterationLevels = 4;
-    repeated Variable transport_context = 5;
 }
 
 message SwimlaneContext {
